folder restructuring and naming convention
